+++ src/lib/components/state/ZeroState.tsx
@@ -1,48 +1,97 @@
 import { Flex, Button, Icon, Text } from "@chakra-ui/react";
+import { useWallet } from "@cosmos-kit/react";
 import { useRouter } from "next/router";
-import { MdOutlineAdd, MdBookmarkBorder, MdSearch } from "react-icons/md";
+import {
+  MdOutlineAdd,
+  MdBookmarkBorder,
+  MdSearch,
+  MdLink,
+} from "react-icons/md";
 
 import { SaveNewContract } from "lib/components/modal/contract";
 import type { Option } from "lib/types";
 
 interface ZeroStateProps {
   list: Option;
-  isReadOnly: boolean;
+  isReadOnly?: boolean;
+  isInstantiatedByMe: boolean;
 }
 
-export const ZeroState = ({ list, isReadOnly }: ZeroStateProps) => {
+const ActionSection = ({
+  isInstantiatedByMe,
+  handleAction,
+  list,
+}: {
+  isInstantiatedByMe: boolean;
+  handleAction?: () => void;
+  list: Option;
+}) => {
+  return isInstantiatedByMe ? (
+    <Button rightIcon={<MdOutlineAdd />} onClick={handleAction}>
+      Deploy New Contract
+    </Button>
+  ) : (
+    <Flex alignItems="center" gap="4" color="text.dark" direction="column">
+      <Flex align="center">
+        Save existing contracts to the list with
+        <SaveNewContract
+          list={list}
+          buttonProps={{
+            variant: "outline-primary",
+            rightIcon: <MdBookmarkBorder />,
+            children: "Save Contract",
+            ml: 2,
+          }}
+        />
+      </Flex>
+      Created contract list and saved contracts are stored in your device only.
+    </Flex>
+  );
+};
+
+/**
+ *
+ * @todo Will be refactored in the next PR
+ */
+
+export const ZeroState = ({
+  list,
+  isReadOnly,
+  isInstantiatedByMe,
+}: ZeroStateProps) => {
   const router = useRouter();
+  const { isWalletConnected, connect } = useWallet();
+
+  if (!isWalletConnected && isInstantiatedByMe) {
+    return (
+      <Flex align="center" color="text.dark" justify="center">
+        <Button
+          variant="outline-primary"
+          rightIcon={<MdLink />}
+          mr={2}
+          onClick={connect}
+        >
+          Connect Wallet
+        </Button>
+        to deploy new contract
+      </Flex>
+    );
+  }
+
   return (
     <Flex alignItems="center" flexDir="column" gap="4">
       <Icon as={MdSearch} color="gray.600" boxSize="16" />
-      <Text color="gray.500">
-        You don’t have any deployed or saved contracts.
+      <Text color="text.dark">
+        {isInstantiatedByMe
+          ? "Your deployed contract through this address will display here"
+          : "You don’t have any saved contracts."}
       </Text>
       {!isReadOnly && (
-        <>
-          <Flex alignItems="center" gap="2">
-            <Text color="gray.500">You can</Text>
-            <Button
-              rightIcon={<MdOutlineAdd />}
-              onClick={() => router.push("/deploy")}
-            >
-              Deploy New Contract
-            </Button>
-          </Flex>
-          <Flex alignItems="center" gap="2">
-            <Text color="gray.500">
-              or save deployed contracts to lists with
-            </Text>
-            <SaveNewContract
-              list={list}
-              buttonProps={{
-                variant: "outline-primary",
-                rightIcon: <MdBookmarkBorder />,
-                children: "Save Contract",
-              }}
-            />
-          </Flex>
-        </>
+        <ActionSection
+          isInstantiatedByMe={isInstantiatedByMe}
+          list={list}
+          handleAction={() => router.push("/deploy")}
+        />
       )}
     </Flex>
   );