+++ src/lib/components/table/EditableCell.tsx
@@ -125,12 +125,7 @@ export const EditableCell = ({
           </Flex>
           {!!tooltip && (
             <Tooltip hasArrow label={tooltip} bg="primary.dark" placement="top">
-              <InfoIcon
-                alignItems="center"
-                color="gray.600"
-                boxSize="4"
-                cursor="pointer"
-              />
+              <InfoIcon color="gray.600" boxSize="14px" cursor="pointer" />
             </Tooltip>
           )}
           {!!onSave && (